feat: show properly form
