+++ frontend/src/components/auth/sign-in-form.tsx
@@ -1,55 +1,11 @@
-import { Controller, useForm } from 'react-hook-form';
-import { ReactElement, useCallback, useState } from 'react';
-import Alert from '@mui/material/Alert';
 import Box from '@mui/material/Box';
-import Button from '@mui/material/Button';
-import { Flag } from '@phosphor-icons/react';
-import FormControl from '@mui/material/FormControl';
-import FormHelperText from '@mui/material/FormHelperText';
-import InputLabel from '@mui/material/InputLabel';
-import OutlinedInput from '@mui/material/OutlinedInput';
-import { authClient } from '@/lib/auth/custom/client';
-import { useUser } from '@/hooks/use-user';
-import { z as zod } from 'zod';
-import { zodResolver } from '@hookform/resolvers/zod';
-
-const schema = zod.object({
-  email: zod.string().email({ message: 'This email address is not valid.' }),
-});
-
-type Values = zod.infer<typeof schema>;
-
-const defaultValues = { email: '' } satisfies Values;
+import FormEmailInput from './form-email-input';
+import FormUsernameInput from './form-username-input';
+import { ReactElement } from 'react';
+import { useFeatures } from '@/hooks/use-features';
 
 export const SignInForm = (): ReactElement => {
-  const { checkSession } = useUser();
-
-  const [isPending, setIsPending] = useState<boolean>(false);
-
-  const {
-    control,
-    handleSubmit,
-    setError,
-    formState: { errors },
-  } = useForm<Values>({ defaultValues, resolver: zodResolver(schema) });
-
-  const onSubmit = useCallback(
-    async (values: Values): Promise<void> => {
-      setIsPending(true);
-
-      const { error } = await authClient.signUp(values);
-
-      if (error) {
-        setError('root', { type: 'server', message: error });
-        setIsPending(false);
-        return;
-      }
-
-      // Refresh the auth state
-      await checkSession?.();
-    },
-    [checkSession, setError],
-  );
+  const { features } = useFeatures();
 
   return (
     <Box
@@ -61,54 +17,7 @@ export const SignInForm = (): ReactElement => {
         alignSelf: 'stretch',
       }}
     >
-      <form onSubmit={handleSubmit(onSubmit)} style={{ width: '100%' }}>
-        <Box
-          sx={{
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'flex-start',
-            alignSelf: 'stretch',
-          }}
-          gap={2}
-        >
-          <Controller
-            control={control}
-            name='email'
-            render={({ field }) => (
-              <FormControl error={Boolean(errors.email)} fullWidth>
-                <InputLabel>What is your email address?</InputLabel>
-                <OutlinedInput {...field} type='email' placeholder='Email' />
-                {errors.email ? (
-                  <FormHelperText>{errors.email.message}</FormHelperText>
-                ) : null}
-              </FormControl>
-            )}
-          />
-          {errors.root ? (
-            <Alert color='error'>{errors.root.message}</Alert>
-          ) : null}
-          <FormHelperText
-            sx={{ color: 'text.primary', whiteSpace: 'pre-wrap' }}
-          >
-            You’ll receive a unique ID code, keep it safe to claim your reward
-            if you win.{'\n'}
-            Your email will not be retained or used for anything other than
-            sending this unique code.
-          </FormHelperText>
-          <Button
-            disabled={isPending}
-            type='submit'
-            variant='contained'
-            fullWidth
-            endIcon={<Flag />}
-            sx={{
-              color: 'text.primary',
-            }}
-          >
-            Join the rebels
-          </Button>
-        </Box>
-      </form>
+      {features?.sendEmail ? <FormEmailInput /> : <FormUsernameInput />}
     </Box>
   );
 };
